+++ src/modules/loadout/index.js
@@ -1,4 +1,5 @@
 export * from '../../bag-shape.js';
 export * from '../../grid-geometry.js';
 export * from '../../backpack-loadout.js';
+export * from '../../artifact-capabilities.js';
 export * from '../../loadout-validation.js';